Add resource_hints to Assests package
