+++ modules/comments.php
@@ -13,6 +13,22 @@
  * @package Jetpack.
  */
 
+use Automattic\Jetpack\Assets;
+
+Assets::add_resource_hint(
+	array(
+		'//jetpack.wordpress.com',
+		'//s0.wp.com',
+		'//s1.wp.com',
+		'//s2.wp.com',
+		'//public-api.wordpress.com',
+		'//0.gravatar.com',
+		'//1.gravatar.com',
+		'//2.gravatar.com',
+	),
+	'dns-prefetch'
+);
+
 /*
  * Add the main commenting system.
  */
@@ -34,16 +50,3 @@ function jetpack_comments_load() {
 
 add_action( 'jetpack_modules_loaded', 'jetpack_comments_load' );
 
-Jetpack::add_resource_hint(
-	array(
-		'//jetpack.wordpress.com',
-		'//s0.wp.com',
-		'//s1.wp.com',
-		'//s2.wp.com',
-		'//public-api.wordpress.com',
-		'//0.gravatar.com',
-		'//1.gravatar.com',
-		'//2.gravatar.com',
-	),
-	'dns-prefetch'
-);
