+++ modules/photon.php
@@ -14,7 +14,7 @@
  * @package Jetpack.
  */
 
-Jetpack::add_resource_hint(
+\Automattic\Jetpack\Assets::add_resource_hint(
 	array(
 		'//i0.wp.com',
 		'//i1.wp.com',
