+++ packages/assets/README.md
@@ -1,8 +1,11 @@
 # Jetpack Asset Management
 
-A package containing functionality to manipulate paths and enqueing async scripts.
+A package containing functionality to improve loading of assets (scripts, etc).
 
-### Usage TODO
+Includes manipulation of paths, enqueuing async scripts, and DNS resource hinting.
+
+### Usage
+`::add_resource_hint( $urls, $type )` -- Adds domains (string or array) to the WordPress' resource hinting. Accepts type of dns-prefetch (default), preconnect, prefetch, or prerender. 
 
 ### Testing
