+++ packages/assets/tests/php/test-assets.php
@@ -179,4 +179,45 @@ class AssetsTest extends TestCase {
 			array( array( 'handle', Assets::get_file_url_for_environment( '/minpath.js', '/path.js' ), array(), '123', true ) )
 		);
 	}
+
+	/**
+	 * Tests the add_resource_hint function.
+	 *
+	 * @author kraftbj
+	 * @covers Assets::add_resource_hint
+	 * @since 8.8.0
+	 */
+	public function test_add_resource_hint_single_string() {
+		/*
+		 * First test is to confirm that the function works when passing a single string with no type.
+		 */
+		$url = '//single-string.example.com';
+		Assets::add_resource_hint( $url );
+		$values = apply_filters( 'wp_resource_hints', array(), 'dns-prefetch' );
+
+		$this->assertContains( $url, $values );
+	}
+
+	/**
+	 * Tests the add_resource_hint function.
+	 *
+	 * @author kraftbj
+	 * @covers Jetpack::add_resource_hint
+	 * @since 8.8.0
+	 */
+	public function test_add_resource_hint_array() {
+		/**
+		 * Next, test an array.
+		 */
+		$urls = array(
+			'//array-1.example.com',
+			'//array-2.example.com',
+		);
+		Assets::add_resource_hint( $urls );
+		$values = apply_filters( 'wp_resource_hints', array(), 'dns-prefetch' );
+
+		$this->assertContains( $urls[0], $values );
+		$this->assertContains( $urls[1], $values );
+
+	}
 }
