+++ phpunit.xml.dist
@@ -1,4 +1,4 @@
-<phpunit bootstrap="tests/php/bootstrap.php" backupGlobals="false" colors="true">
+<phpunit bootstrap="tests/php/bootstrap.php" backupGlobals="false" colors="true" convertDeprecationsToExceptions="true">
 	<php>
 		<const name="PHPUNIT_JETPACK_TESTSUITE" value="true"/>
 	</php>
